backend/requirement.txt
openai>=1.0.0
moviepy==1.0.3
pillow
requests
soundfile
tqdm
numpy
dashscope
imageio[opencv]
opencv-python
flask==2.2.2
werkzeug==2.2.2
redis
tweepy
schedule
ormsgpack
typing_extensions
python-dotenv
aptos-sdk
Flask[async]
python-dateutil
pytz